requirements.txt
pandas>=1.0.0
plotly>=5.0.0
requests>=2.32.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
flask>=2.0.0
azure-functions>=1.0.0
